Added scenario for node deactivation test
